// src/modules/ai/core/types/languages.ts
export enum Languages {
  Arabic_Egyptian = "ar-EG",
  German_Germany = "de-DE",
  English_US = "en-US",
  Spanish_US = "es-US",
  French_France = "fr-FR",
  Hindi_India = "hi-IN",
  Indonesian_Indonesia = "id-ID",
  Italian_Italy = "it-IT",
  Japanese_Japan = "ja-JP",
  Korean_Korea = "ko-KR",
  Portuguese_Brazil = "pt-BR",
  Russian_Russia = "ru-RU",
  Dutch_Netherlands = "nl-NL",
  Polish_Poland = "pl-PL",
  Thai_Thailand = "th-TH",
  Turkish_Turkey = "tr-TR",
  Vietnamese_Vietnam = "vi-VN",
  Romanian_Romania = "ro-RO",
  Ukrainian_Ukraine = "uk-UA",
  Bengali_Bangladesh = "bn-BD",
  English_India = "en-IN",
  Marathi_India = "mr-IN",
  Tamil_India = "ta-IN",
  Telugu_India = "te-IN",
}

export enum Voices {
  Zephyr = "Zephyr",
  Puck = "Puck",
  Charon = "Charon",
  Kore = "Kore",
  Fenrir = "Fenrir",
  Leda = "Leda",
  Orus = "Orus",
  Aoede = "Aoede",
  Callirrhoe = "Callirrhoe",
  Autonoe = "Autonoe",
  Enceladus = "Enceladus",
  Iapetus = "Iapetus",
  Umbriel = "Umbriel",
  Algieba = "Algieba",
  Despina = "Despina",
  Erinome = "Erinome",
  Algenib = "Algenib",
  Rasalgethi = "Rasalgethi",
  Laomedeia = "Laomedeia",
  Achernar = "Achernar",
  Alnilam = "Alnilam",
  Schedar = "Schedar",
  Gacrux = "Gacrux",
  Pulcherrima = "Pulcherrima",
  Achird = "Achird",
  Zubenelgenubi = "Zubenelgenubi",
  Vindemiatrix = "Vindemiatrix",
  Sadachbia = "Sadachbia",
  Sadaltager = "Sadaltager",
  Sulafat = "Sulafat",
}